Move VisuallyHidden files up a dir
