+++ packages/components/src/a11y/index.ts
@@ -1,2 +1,2 @@
 export * from "./SkipLink"
-export * from "./VisuallyHidden/VisuallyHidden"
+export * from "./VisuallyHidden"
